feat: unify column matching helpers
