+++ tests/test_ai_suggestions.py
@@ -1,10 +1,10 @@
 import pytest
-from services.ai_suggestions import generate_column_suggestions
+from services.data_enhancer import get_ai_column_suggestions
 
 
 def test_basic_suggestions():
     columns = ["Time", "Person ID", "Token ID", "Door Name", "Result", "Other"]
-    suggestions = generate_column_suggestions(columns)
+    suggestions = get_ai_column_suggestions(columns)
     assert suggestions["Time"]["field"] == "timestamp"
     assert suggestions["Person ID"]["field"] == "person_id"
     assert suggestions["Token ID"]["field"] == "token_id"
@@ -14,6 +14,6 @@ def test_basic_suggestions():
 
 
 def test_unknown_column_returns_empty():
-    suggestions = generate_column_suggestions(["Mystery"])
+    suggestions = get_ai_column_suggestions(["Mystery"])
     assert suggestions["Mystery"]["field"] == ""
     assert suggestions["Mystery"]["confidence"] == 0.0
